Scale ShardScheduler
As we scale the number of nodes, a single ShardScheduler thread becomes
a bottleneck since it needs to iterate through many replicas. We now
scale it accordingly, which is enough to demonstrate constant time
under reasonable cluster sizes. Sadly Python isn't the best tool out
there for threaded stuff, so the invariant breaks under crazy values.
But not a boomer.